+++ assets/javascripts/kitten/components/form/text-input-with-button/index.js
@@ -6,6 +6,7 @@ import TYPOGRAPHY from '../../../constants/typography-config'
 import { pxToRem, stepToRem } from '../../../helpers/utils/typography'
 import COLORS from '../../../constants/colors-config'
 import { modifierStyles } from '../../../components/buttons/button/helpers/modifier-styles'
+import deprecated from 'prop-types-extra/lib/deprecated'
 
 const StyledTextInputWithButton = styled.div`
   display: flex;
@@ -61,7 +62,13 @@ export class TextInputWithButton extends PureComponent {
     disabled: PropTypes.bool,
     tiny: PropTypes.bool,
     modifier: PropTypes.string,
-    value: PropTypes.string,
+    inputValue: PropTypes.string,
+    // DEPRECATED: do not use prop `value`. Use `buttonValue` instead.
+    value: deprecated(
+      PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
+      'Use `buttonValue` instead.',
+    ),
+    buttonValue: PropTypes.oneOfType([PropTypes.string, PropTypes.node]),
   }
 
   static defaultProps = {
@@ -69,7 +76,7 @@ export class TextInputWithButton extends PureComponent {
     error: false,
     disabled: false,
     tiny: false,
-    value: 'Button',
+    buttonValue: 'Button',
     modifier: 'beryllium',
   }
 
@@ -80,8 +87,10 @@ export class TextInputWithButton extends PureComponent {
       disabled,
       tiny,
       value,
+      buttonValue,
       modifier,
       buttonProps,
+      inputValue,
       ...others
     } = this.props
 
@@ -93,6 +102,7 @@ export class TextInputWithButton extends PureComponent {
           error={error}
           disabled={disabled}
           tiny={tiny}
+          value={inputValue}
         />
         <StyledButton
           type="button"
@@ -102,7 +112,7 @@ export class TextInputWithButton extends PureComponent {
           disabled={disabled}
           {...buttonProps}
         >
-          {value}
+          {value || buttonValue}
         </StyledButton>
       </StyledTextInputWithButton>
     )