+++ assets/javascripts/kitten/components/form/text-input-with-button/stories.js
@@ -3,6 +3,7 @@ import { storiesOf } from '@storybook/react'
 import { withInfo } from '@storybook/addon-info'
 import { withKnobs, boolean, text } from '@storybook/addon-knobs'
 import { Grid, GridCol } from '../../../components/grid/grid'
+import { Loader } from '../../../components/loaders/loader'
 import { TextInputWithButton } from './index'
 
 const info = {
@@ -37,6 +38,10 @@ const info = {
   `,
 }
 
+const handleChange = e => {
+  console.log('e', e)
+}
+
 storiesOf('Form/TextInput', module)
   .addDecorator(withKnobs)
   .addDecorator(withInfo)
@@ -51,11 +56,13 @@ storiesOf('Form/TextInput', module)
               error={boolean('Error', false)}
               disabled={boolean('Disabled', false)}
               tiny={boolean('Tiny', false)}
-              value={text('Button', 'Button')}
+              value={<Loader />}
+              inputValue={text('inputValue', 'Une valeur')}
               placeholder={text(
                 'Les props sont transmises',
                 'Les props sont transmises',
               )}
+              onChange={handleChange}
             />
           </GridCol>
         </Grid>